MovieReel Cache
Cache implmementation for MovieReel to store data on device. This will
use Room as an abstraction layer over SQlite and  Dao to handle the SQL
queries for CRUD operation on data in the databases.

Mapper(s) will be used to transform the data from Cache Models to Data
Entities and vice versa when there is communication between layers.

Preference Helper is used to handle storage of constants.